Resolve 3 conflicts:
- Dockerfile: keep syntax directive + master comment
- .dockerignore: keep expanded ignore list
- CLAUDE.md: keep v2.0 format

+++ EasyStock.Application/UseCases/ReporEstoque/ReporEstoqueUseCase.cs
@@ -26,8 +26,8 @@ namespace EasyStock.Application.UseCases.ReporEstoque
     public sealed record ReporEstoqueResult(
         Guid ItemEstoqueId,
         Guid MovimentacaoId,
-        int QuantidadeAnterior,
-        int QuantidadeAtual);
+        decimal QuantidadeAnterior,
+        decimal QuantidadeAtual);
 
     public class ReporEstoqueUseCase(
         IProdutoRepository produtoRepository,
