+++ EasyStock.Domain/Events/SaidaEstoqueRegistrada.cs
@@ -2,5 +2,5 @@ using System;
 
 namespace EasyStock.Domain.Events
 {
-    public sealed record SaidaEstoqueRegistrada(Guid EventoId, DateTime OcorridoEm, Guid ItemEstoqueId, Guid ProdutoId, Guid EmpresaId, int Quantidade, string? Motivo) : DomainEvent(EventoId, OcorridoEm);
+    public sealed record SaidaEstoqueRegistrada(Guid EventoId, DateTime OcorridoEm, Guid ItemEstoqueId, Guid ProdutoId, Guid EmpresaId, decimal Quantidade, string? Motivo) : DomainEvent(EventoId, OcorridoEm);
 }
